Extract git-specific code to its own module
